aioradio/requirements.txt
aioboto3==8.0.5
aiobotocore==1.1.2
aiojobs==0.2.2
aioredis==1.3.1
ddtrace==0.42.0
fakeredis==1.4.3
flask==1.1.2
httpx==0.15.5
mandrill==1.0.59
moto==1.3.16
orjson==3.4.0
pre-commit==2.7.1
pylint==2.6.0
pyodbc==4.0.30
pysmb==1.2.2
pytest==6.1.1
pytest-asyncio==0.14.0
pytest-cov==2.10.1
python-json-logger==2.0.0
